client-repo-docker/requirements.txt
requests==2.26.0
python-dotenv==0.19.2
selenium==3.14.0
boto3
six
rsa
s3transfer==0.5.1